Refactored by moving files
- Moved ContainerRegistryHandler and TileEntityRegistryHandler
into registry_handlers package.
