+++ src/main/java/com/eonzenx/mcppmod/util/registry_handlers/ContainerRegistryHandler.java
@@ -1,4 +1,4 @@
-package com.eonzenx.mcppmod.util;
+package com.eonzenx.mcppmod.util.registry_handlers;
 
 import com.eonzenx.mcppmod.MCPPMod;
 import com.eonzenx.mcppmod.objects.containers.SoupPotBlockContainer;
